plugins/taskmanager-agent/pyproject.toml
[project]
name = "taskmanager-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["httpx>=0.27", "click>=8.1", "pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-httpx>=0.30"]
